code gen optimized
Added a field spec modifier design to the framework.

Handling annotations for the fields annotated with Convert or implemented from BaseEnum is now supported.

It rewrites complex types to specific types, such as date types to strings or long types to strings.

Once the fields in response or request models are converted, the rest of the data mapping work falls on the mappers in MapStruct.

Signed-off-by: hupeng <hup_dev@outlook.com>